refactor: Rename ObjectPool
